Fix potential deadlock in onServiceConnected method.
`onServiceConnected` synchronizes on `mLock` which is a local lock declared
in `BluetoothDeviceManager`. Inside that synchronization block it calls
`InCallController#bindToBTService` which itself locks on the Telecom sync
root.  Ideally we shouldn't hold both locks at the same time.

Renamed the local `mLock` to `mDtDeviceManagerLock` to make it more clear
that the `mLock` isn't the Telecom sync root.  Also wrapped the call to
`CallsManager` in a sync on Telecom's lock since `CallsManager#hasAnyCalls`
is not synchronized itself.

Flag: EXEMPT BUG_FIX
Fixes: 442941161
Test: Re-run Telecom unit and CTS tests
Change-Id: I63d1456b38cb15575441715da02ee3b6b1581024