Prevent inactive reroute on SCO disconnect
Ensures that when we receive indication from the audio fwk that the
communication device was updated out of SCO, that we do not try to
recalculate the audio routing if it's in an ianctive state. The BTHS is
still paired to the device so we can remain in the route while we're out
of the call. The communication device updates received outside of a call
were due to BT stack using the set/clear communication device APIs. We
were rerouting audio to earpiece and ended up starting the call on
earpiece since we no longer recalculate audio routing when receiving an
active focus switch (due to preserve_call_audio_routing flag).

Bug: 442657101
Flag: EXEMPT bug fix
Test: atest CallAudioRouteControllerTest
Change-Id: Ideb10899465d77d50498ae3c52fc413ecdb6e861