Auto unhold on call failure
For calls made on different phone accounts, we may try to hold the
ongoing call and if the second call fails to be set up, the bg call will
not move back to the foreground. Ensure that we auto-unhold if the
disconnect is due to an error or cancellation reason.

Bug: 441531098
Flag: com.android.server.telecom.flags.auto_unhold_on_call_fail
Test: atest CallSequencingBasicCallTest
Change-Id: Ifd92d1e3690fbcffa4f8581240196a27c99abb01