Recalculate routing on fg call updates
Ensure that we recalculate the route when the fg call is updated. This
ensures that if, for instance, we have an ongoing video call (on
speaker) and place a new PSTN call that the call is placed on earpiece.
This change also ensures that we still preserve the audio routing when
we receive active focus switches as to not cause regressions with the
changes introduced by the preserve_call_audio_routing flag.

Bug: 443356737
Flag: EXEMPT bug fix
Test: atest CallAudioRouteControllerTest
Test: Manual to verify that putting a normal call on speaker + hold +
unhold leaves the call in speaker
Test: WhatsApp video call + normal PSTN to verify that the latter call
is routed to earpiece. Switching between the calls after that to ensure
that audio routing is mainting on the appropriate route.

Change-Id: I0a52bb2bbcc476fbdbcecd65f665bb33613e22c1