Always set communication device at start of call
Ensure that we always set the communication device at the start of the
call even if the destination route is already the current communication
device being reported by the audio fwk. Once Telecom becomes the mode
owner, it's possible that the current communication device is cleared by
the previous mode owner so Telecom needs to request the device again. We
can skip adding the pending message check in the case that Telecom does
not receive the onCommunicationDeviceChanged callback update.

Bug: 444111674
Flag: EXEMPT BUGFIX
Test: 2 new unit tests in CallAudioRouteControllerTest
Change-Id: Ia27307fd84f9e5a127e1b211484414eab4c49cb4