Local Voicemail implementation Part 3
0. Remove TelecomManager LVM disabled constant.
1. Ensuring call state change verification succeeds when we answer and go
to local voicemail state.
2. Call log manager logging improvements.
3. Add string versions of new call states for readability in logs.
4. Ensure that when a call is ended during local voicemail that it is
marked as missed; this ensures it generates a missed call notification and
gets marked as missed in the call log.
5. Handle removing a call from the InCallController when local voicemail
starts.  Refactored the logic which onExternalCallStateChanged used to do
the add and remove and reused it for local voicemail cases, but with an
optional parameter to specify a disconnect cause.
6. Fix crash in LocalVoicemailController where a call goes away before
the local voicemail service finishes binding.
7. Update ParcelableCallUtils to add a prototype that allows passing in
an override disconnect cause.
8. Added shell command to change the local voicemail timeout for all phone
accounts for test purposes.

Flag: com.android.server.telecom.flags.local_voicemail
Test: CTS tests added
Bug: 405460896
Bug: 394367444
Bug: 405460896
Change-Id: I5de09e19c489844172d1f2802a7e7a50ea3f68c4