Clean up VOIP call log test app
Allow an option for the user to clear the VOIP call logs from the test
dialer app. Also, whenever the VOIP app main activity is shut down, also
clear the VOIP call log entries so that we do not have any stale entries
that result in callbacks not working (uuids are stored in memory).

Bug: 445941695
Flag: EXEMPT TEST_ONLY
Test: test w/ transactional test app
Change-Id: I519d458fe10c88ef93be8e3bf97575b2690cf683