Update focus call on set active
Ensure that the focus call is indeed the call to be marked as active. It
is possible that a call is answered directly without user intervention
via the incall UI (i.e. answering via ATA cmd). We should ensure that we
update the focus call if it is not reflected.

Bug: 445832417
Flag: com.android.server.telecom.flags.request_focus_for_set_active
Test: atest TelecomUnitTests
Change-Id: I9c65422e2a95954ac37758085ba6575ab6fe2524