Skip reroute for BT_DEVICE_REMOVED
If there's a pending audio route that is not on the device being
disconnected, then we should not try to recalculate the audio routing.
The current logic only accounts for the current route. However, if
there's a switch to speaker that's pending and we remove the BT device,
then the audio route will be recalculated to route into earpiece. This
overrides the speaker route request.

Bug: 446241982
Flag: EXEMPT BUGFIX
Test: atest CallAudioRouteControllerTest
Change-Id: I9e4a4d746f3035659156fe2c7f4a7f318a80b8a1