Ensure that calls always have a creation time.
Although AOSP's constructor for `com.android.server.telecom.Call` always
sets the `mCreationTimeMillis`, there appear to be some devices in market
where the resulting `android.telecom.Call.Details` passed to the dialer
has a zero call creation time.  Although we've updated the CTS test it does
not guarantee that the call creation time will be > 0 in all possible code
paths.  Since creation time is always accessed through the getter, we add
a correctness check in the getter to ensure that the creation time is
always set.  The getter is NOT called from the constructor.

Test: Tested by CTS CallDetailsTest.
Fixes: 448734781
Flag: NONE BUGFIX
Change-Id: I96b515e8b3277e83a144a595203f3065411a9ea7