Track hearing aid pair in AudioRoute
Enusre that we track the hearing aid pair instead of ignoring the device
when it gets connected. This will ensure that if the first connected
device is removed that we are still keeping track of the connected
second device. This will also allow us to use both addresses when we try
to find the device via AudioManager#getAvailableCommunicationDevices
when we're setting the communication device for the route.

Bug: 390523404
Flag: com.android.server.telecom.flags.hearing_aid_pair_fix
Test: atest CallAudioRouteControllerTest
Change-Id: Ia47a8f98ddd03da5fca47c5884e53b5ade63fdea