Store previous active device in cache
Stores the previously active device in the route controller cache to
properly handle the case when we receive BT_ACTIVE_DEVICE_GONE. The
current implementation checks the cache directly but the active device
update occurs immediately in BluetoothStateReceiver before we send the
BT_ACTIVE_DEVICE_GONE msg to the controller, which leads to a stale
update. This was causing issues with hearing aids that were in the
charger and the audio routing to LE despite Telecom receiving the
inactive device update.

Bug: 450144098
Flag: EXEMPT BUGFIX
Test: atest CallAudioRouteControllerTest
Change-Id: I7d9fe1777858734d3aec3ee8e7a5a44a564b82b0