Revise hearing aid pair fix logic
The original issue was raised as a result of a NPE but upon closer
investigation, this unrevealed a deeper flaw with the current logic. The
initial implementation assumed that we could just modify the existing
audio route without any additional updates to the other caches we track
(i.e. available routes, call supported routes, bt route to bt device
mapping). However, modifying the routes causes the calculated hash to be
different and because those routes were being stored in the
aforementioned caches, the newly modified routes were not getting
resolved in the cache. Instead, we have to ensure that we remove the
existing references from the caches and replace them with the new
modified routes.

Bug: 451677137
Flag: com.android.server.telecom.flags.hearing_aid_pair_fix
Test: atest CallAudioRouteControllerTest
Test: (1) Connect one bud and then the next. Make a call to verify audio
routed via HA. Disconnect the active bud and verify call audio is still
routed to the HA and is reflected in the UI. Reconnect the bud and
verify audio is routed to HA. Disconnect the same bud to verify audio is
still routed to HA. In between the disconnect/reconnects, I verified
that switching between HA -> speaker and HA -> earpiece and back to HA
was WAI. I also ran a similar test case for WhatsApp

Change-Id: I93980e8618906f39fb10278aa7f191c91877b07a