Integrate new InCallService#onCallEndpointRequested API in Telecom
This CL integrates a new InCallService#onCallEndpointRequested API in Telecom that will be invoked for all other ICS’s whenever ICA#setAudioRoute or ICA#requestCallEndpointChange are invoked. It is intended so that a non-UI ICS such as a companion App can be aware of audio route changes initiated by the user through the Dialer application.

Test: atest ThirdPartyInCallServiceAppOpsPermissionTest
Bug: 426100640
Flag: android.telecom.flags.call_endpoint_requested_api
Change-Id: I7928c5f73ea985297a81110dc53ab02e762662e3