Add test for getCreationTimeMillis defensive check
Adds a unit test to verify the defensive check within the
`getCreationTimeMillis` method. This test ensures that if a `Call`
object is created with a creation time of 0, the getter will correctly
assign a valid, non-zero timestamp on its first invocation.



This CL is generated by AI, please check the comments for more details.

Test: ATP tests passed http://go/forrest-run/L54500030029272674
Bug: 431235865
Flag: TEST_ONLY



Change-Id: Ie6bd93795ed64d847a79ba463dea00c27624ce3b