Only send focus switch if cached call exists
For delaying the routing calculation to waith for the BT ICS binding,
ensure that we only send the corresponding focus switches if the
associated cached calls still exist. In the case of an external call
state change, the call may become external while the BT ICS future is
pending and we will send the focus switch even though the mode should be
normal.

Bug: 454420574
Flag: EXEMPT BUGFIX
Test: atest ExternalCallTest
Change-Id: Ic4dee5984e70fa374fd309bba1f1e1dbd7abcaeb