Only update route when fg is present
Ensure that the route is only updated if there's a valid fg call. When a call is disconnected. the fg call is not defined but we may still be playing a disconnect tone. In this case, a call on speaker will end up playing the disconnect tone over earpiece instead.

Explained in the code, but we should ideally only update routing when the fg change results in a call with a different phone account from the previous fg call. This will be resolved as part of the pending state removal improvements that will be done in 26Q2.

Bug: 452783026
Flag: EXEMPT BUGFIX
Test: Manual
Change-Id: I75acea5c16a91776ce98b200ad871a7776710609