Retrieve the main user and associated context in BlockCheckerFilter.
This CL ensures the context associated with the correct  main user is being used when we check if a number is blocked in BlockCheckerFilter#startFilterLookup. This CL also adds a flag to guard this change and it updates the Telecom dumpsys to include the dump of Telecom Module bug fix flags.

Test: atest BlockCheckerFilterTest
Flag: com.android.internal.telecom.flags.get_main_user_for_block_checker
Bug: 445311509
Change-Id: I275d45af08ca388ab1fcf2388c9675c0dca6b48f