Ensure ACTION_PACKAGE_CHANGED receiver is registered after unbind.
We use a Context for the "all users" context to register an
ACTION_PACKAGE_CHANGED receiver which should live for the duration of all
ongoing calls.  It turns out we were trying to UNREGISTER from the wrong
context, so we never actually did unregister.

The unregister had a try/catch that was masking this issue entirely.
It looks like that was added because `unbindFromServices` is also called
from a telecom shell command to forcefully cleanup all remaining calls, and
it is entirely possible there are just no calls and no bindings.

To fix that I added an AtomicBoolean to guard the register and unregister
so that we track registration and unregistration state.

Test: Manual test with log inspection.
Flag: EXEMPT BUGFIX
Fixes: 460002064
Change-Id: I0955969333bba08eb88a5e4ca2b8918dc5bb6fb9