Skip ICS update when focus is abandoned
When there are no calls (or all of them are disconnected) and we don't
hold audio focus, we should refrain from sending audio style updates to
the ICS. At that point, we may encounter a race with when we abandon
focus and when the ICS are unbound. In some cases, it's possible for the
AF to process the abandon focus request much earlier and another app may
change the audio route once they gain focus. If the incall screen is
still visible, the user may see those route changes.

Bug: 459266282
Flag: com.android.internal.telecom.flags.skip_audio_update_to_ics
Test: Manual testing with two call scenario and switching audio routes
Change-Id: I00abd02ff47eca3208303b98c0725da564d8f064