Add tests for VoNR call feature logging
Add unit tests for CallLogManager to verify the logging of Voice over New Radio (VoNR) calls.

These tests confirm that the `Calls.FEATURES_VONR` flag is correctly added to the call log entry for a VoNR call when the feature is enabled. They also ensure the flag is not added when the corresponding feature flag is disabled.



This CL is generated by AI, please check the comments for more details.

Test: ATP tests passed http://go/forrest-run/L58100030038704597
Bug: 431235865 (Umbrella)
Flag: TEST_ONLY



Change-Id: Ibf745fb35ee18d94b0b3269f8b5195e29375ce07