when stopFGSDelegation is called, remove the service connection
An external dev reported that when an app has fgs-delegation
and is force stopped, the next call is not re-granted fgs
delegation. This is a result of the map not being cleaned
up in the stopFGSDelegation function.

The fix is to remove the handle from the mapping of
services being tracked since we are calling stopFgs.

In this case, since the service has already been stopped,
the AM does not call onDisconnected and we never
removed the service in that callback. This lead to a stale
state.

Flag: EXEMPT BUGFIX
Fix: 447185999
Test: manual - start a call, kill app, then add another call
     expect: new FGS delegation
Change-Id: I25ce6df5484db825e607f0403dfddab1b38fc654