Fix flaky CallAudioRouteController tests
Add extra waiting logic to confirm that the call audio state was changed
instead of directly checking the current route. It's possible with the
timing induced by the tests for the verifications to not be true at the
time of assertion. Instead, we can use a timeout to verify the audio
state was updated to the expected route.

Bug: 469409216
Flag: EXEMPT BUGFIX
Test: atest CallAudioRouteControllerTest
Change-Id: I0b0701d224412b07c8d9c3c4a39f619787d8c13f