Fix: Resolve ClassNotFoundException crash in RespondViaSmsSettings
This change fixes a crash that occurs when launching the RespondViaSmsSettings activity for editing quick call responses.

After the decoupling of the TelecomUi module, launching the "Quick responses" screen causes a ClassNotFoundException because the activity attempts to inflate a layout containing the MultiLineTitleEditTextPreference class.
This class is internal to the Telecomm module and is no longer accessible to the sandboxed TelecomUi app after the TelecomLib dependency was removed

Bug: 470064876
Test: manual
Flag: EXEMPT PURE_REFACTOR
Change-Id: I5187d6a1268292683bf31513550a2593554cc9b3