Cleanup flag enable_ringtone_haptics_customization in Telecom.
This is a mainline cleanup that resolves a dependency build error by
ensuring that the media.audio flag enable_ringtone_haptics_customization
is enabled in Telecom's Ringer.java. Since the flag is a media audio
flag that is not exported (and cannot be exported since it doesn't guard
any APIs), this flag cannot be referenced in Telecom so our options are
either #1 treat this flag as if it is enabled in Telecom or #2 compltely
rollback the flag and remove it from Telecom.

Bug: 469913991
Flag: EXEMPT FLAG_REMOVAL
Test: atest TelecomUnitTests
Change-Id: I2760f3b6e362438baaad5ed6a0200df652282ed5