Replace hidden AudioManager API in Ringer for Mainline compatibility
The method `AudioManager.shouldNotificationSoundPlay()` is a hidden API
and cannot be used in the Telecom Mainline module. This CL replaces it
with a functional equivalent using public APIs to check if the ringtone
should be audible.

Until the SystemApi is added, we will revert back to the old behavior

Flag: EXEMPT PURE_REFACTOR
Bug: 469913991 (root mainline bug)
Bug: 470401391 (bug to switch back to shouldNotificationSoundPlay)
Test: none / presubmits
Change-Id: I15b18c8bdb4769cf55145ab92713d0481dd4dc7c